perf(#zimic): optimize fetch client type checking
