+++ packages/zimic/src/fetch/factory.ts
@@ -2,10 +2,10 @@ import { HttpSchema } from '@/http';
 import { ConvertToStrictHttpSchema } from '@/http/types/schema';
 
 import FetchClient from './FetchClient';
-import { FetchClientOptions, Fetch as PublicFetch } from './types/public';
+import { FetchOptions, Fetch as PublicFetch } from './types/public';
 
 function createFetch<Schema extends HttpSchema>(
-  options: FetchClientOptions<ConvertToStrictHttpSchema<Schema>>,
+  options: FetchOptions<ConvertToStrictHttpSchema<Schema>>,
 ): PublicFetch<ConvertToStrictHttpSchema<Schema>> {
   const { fetch } = new FetchClient<ConvertToStrictHttpSchema<Schema>>(options);
   return fetch;
