+++ packages/zimic/src/fetch/types/public.ts
@@ -3,7 +3,7 @@ import { LiteralHttpSchemaPathFromNonLiteral } from '@/http/types/schema';
 import { Default, PossiblePromise } from '@/types/utils';
 
 import FetchResponseError from '../errors/FetchResponseError';
-import { FetchRequest, FetchRequestInit, FetchResponse } from './requests';
+import { FetchRequest, FetchRequestConstructor, FetchRequestInit, FetchResponse } from './requests';
 
 export type FetchInput<
   Schema extends HttpSchema,
@@ -46,23 +46,11 @@ export interface FetchFunction<Schema extends HttpSchema> {
   >;
 }
 
-export interface FetchClientOptions<Schema extends HttpSchema> extends FetchRequestInit.Defaults {
+export interface FetchOptions<Schema extends HttpSchema> extends FetchRequestInit.Defaults {
   onRequest?: (this: FetchClient<Schema>, request: FetchRequest.Loose) => PossiblePromise<FetchRequest.Loose>;
 
   onResponse?: (this: FetchClient<Schema>, response: FetchResponse.Loose) => PossiblePromise<FetchResponse.Loose>;
 }
-
-export type FetchRequestConstructor<Schema extends HttpSchema> = new <
-  Path extends HttpSchemaPath.NonLiteral<Schema, Method>,
-  Method extends HttpSchemaMethod<Schema>,
->(
-  input: FetchInput<Schema, Path, Method>,
-  init: FetchRequestInit<Schema, LiteralHttpSchemaPathFromNonLiteral<Schema, Method, Path>, Method>,
-) => FetchRequest<
-  LiteralHttpSchemaPathFromNonLiteral<Schema, Method, Path>,
-  Method,
-  Default<Schema[LiteralHttpSchemaPathFromNonLiteral<Schema, Method, Path>][Method]>
->;
 
 export interface FetchClient<Schema extends HttpSchema> {
   defaults: FetchRequestInit.Defaults;
