+++ packages/zimic/src/interceptor/http/interceptor/__tests__/shared/bodies/blob.ts
@@ -61,7 +61,8 @@ export async function declareBlobBodyHttpInterceptorTests(options: RuntimeShared
   let baseURL: URL;
   let interceptorOptions: HttpInterceptorOptions;
 
-  const Handler = options.type === 'local' ? LocalHttpRequestHandler : RemoteHttpRequestHandler;
+  const Handler: typeof LocalHttpRequestHandler | typeof RemoteHttpRequestHandler =
+    options.type === 'local' ? LocalHttpRequestHandler : RemoteHttpRequestHandler;
 
   beforeEach(() => {
     baseURL = getBaseURL();
