+++ packages/zimic/tsconfig.json
@@ -5,7 +5,6 @@
     "target": "esnext",
     "allowSyntheticDefaultImports": true,
     "tsBuildInfoFile": "tsconfig.tsbuildinfo",
-    "allowUnreachableCode": false,
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"],
